requirements.txt
fastapi
pydantic
python-dotenv
vapi-server-sdk